requirements.txt
fastapi
python-multipart
langdetect
openai
openai-whisper
gTTS
uvicorn
ffmpeg-python
python-dotenv
torch --index-url https://download.pytorch.org/whl/cpu
